Обновлено именование. Удален лишний черновой файл
